third_party/WebKit/LayoutTests/http/tests/security/contentSecurityPolicy/inline-script-blocked-expected.txt

CONSOLE ERROR: line 9: Refused to execute inline script because it violates the following Content Security Policy directive: "script-src http://127.0.0.1:*".

CONSOLE ERROR: Refused to execute inline script because it violates the following Content Security Policy directive: "script-src http://127.0.0.1:*".

This test passes if it doesn't alert fail.
